requirements.txt
torch>=2.1.0
transformers>=4.45.0
peft>=0.13.0
datasets>=2.20.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
pyyaml>=6.0
huggingface_hub>=0.25.0
openai>=1.40.0
tqdm>=4.66.0
langdetect>=1.0.9
nltk>=3.8.0
immutabledict>=4.2.0
absl-py>=2.0.0
tenacity>=8.2.0
sentencepiece>=0.2.0
protobuf>=4.25.0
